Update readme.md
Myschoolgateway(https://myschoolgateway.com/) is a universal school management system I solely developed in 2018. It is been used by 13 secondary schools.
It is fully customizable to any school standard and it has different access level for school teacher, principal, administrator and students. This release note is focused mainly on enhancements on school staff dashboard. The overall release notes including the headers and overview is more than 250 words as instructed but the main note on enhancement on the software is below 250 words.